Finalização do projeto

+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\ProductRequest;
+use App\Http\Requests\UpdateProductRequest;
 
 class ProductController extends Controller
 {
@@ -14,7 +15,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::select('id','name','arquivo')->paginate(3);
+        $products = Product::select('id','name','arquivo', 'author')->paginate(3);
         return view('dashboard', compact('products'));
     }
 
@@ -67,15 +68,54 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        //
+        if(auth()->id() !== $product->author){
+            return redirect()->route('dashboard')->with('error', 'Você não tem permissão para alterar o produto');
+        }
+
+        $user = auth()->id();
+        return view('products.edit', compact('product','user'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(ProductRequest $request, Product $product)
-    {
-        //
+    public function update(UpdateProductRequest $request, Product $product)
+    {   
+        if(auth()->id() !== $product->author){
+            return redirect()->route('dashboard')->with('error', 'Você não tem permissão para alterar o produto');
+        }
+
+        if($request->hasFile('arquivo')){
+            Storage::disk('public')->delete('images/'.$product->arquivo);
+
+            $archiveName = $request->file('arquivo')->store('images', 'public');
+            $archiveName = explode('/',$archiveName)[1];
+
+
+            $update = $product->update([
+                'name' => $request['name'],
+                'price' => $request['price'],
+                'weight' => $request['weight'],
+                'weight' => $request['weight'],
+                'arquivo' => $archiveName,
+                'description' => $request['description']
+            ]);
+        }else{
+            $update = $product->update([
+                'name' => $request['name'],
+                'price' => $request['price'],
+                'weight' => $request['weight'],
+                'weight' => $request['weight'],
+                'description' => $request['description']
+            ]);
+        }
+
+        if($update){
+            return redirect()->route('products.index')->with('success', 'Produto atualizado com sucesso');
+        }else{
+            return redirect()->route('products.edit')->with('error', 'Não foi possível atualizar o produto');
+        }
+        
     }
 
     /**
@@ -83,6 +123,10 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {   
+        if(auth()->id() !== $product->author){
+            return redirect()->route('dashboard')->with('error', 'Você não tem permissão para apagar o produto');
+        }
+
         $deletedProduct = $product->delete($product->id);
 
         Storage::disk('public')->delete('images/'.$product->arquivo);
